tortoise-finder/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tortoise-finder"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.7.4",
  "python-multipart==0.0.9",
  "typer==0.12.3",
  "redis==5.0.7",
  "rq==1.16.2",
  "boto3==1.34.131",
  "minio==7.2.7",
  "mlflow==2.14.3",
  "pandas==2.2.2",
  "pyarrow>=4.0.0,<16",
  # "geopandas==0.14.4",  # Temporarily disabled for quick setup
  # "shapely==2.0.4",     # Temporarily disabled for quick setup  
  # "pyproj==3.6.1",      # Temporarily disabled for quick setup
  "pillow==10.3.0",
  "tqdm==4.66.4",
  "gradio==4.41.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "app_ui*", "pipeline*", "storage*", "models*", "cli*"]

[tool.typer]
pretty_exceptions_enable = true
